Optimizacion de todos los EndPoints y manejo de todas las exceptions en sus respectivos Servicios.

+++ src/main/java/com/api/libraryproject/service/BooksService.java
@@ -2,8 +2,13 @@ package com.api.libraryproject.service;
 
 import com.api.libraryproject.dto.BooksDto;
 import com.api.libraryproject.entity.BooksEntity;
+import com.api.libraryproject.entity.UsersEntity;
+import com.api.libraryproject.exceptions.BooksException;
 import com.api.libraryproject.repository.BooksRepository;
+import com.api.libraryproject.repository.LibraryRepository;
+import com.api.libraryproject.util.Role;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import com.api.libraryproject.util.BookStatus;
 
@@ -14,6 +19,8 @@ public class BooksService {
 
     @Autowired
     private BooksRepository booksRepository;
+    @Autowired
+    private LibraryRepository libraryRepository;
 
     public List<BooksDto> getAllBooks() {
         return booksRepository.findAll().stream()
@@ -43,7 +50,14 @@ public class BooksService {
         return books;
     }
 
-    public BooksDto addBook(BooksDto bookDto) {
+    public BooksDto addBook(BooksDto bookDto, String userEmail) {
+        UsersEntity user = libraryRepository.findByEmail(userEmail)
+                .orElseThrow(() -> new UsernameNotFoundException("No se encontro al usuario con el correo: " + userEmail));
+
+        if(user.getRole() != Role.ADMIN){
+            throw new BooksException("Solo los empleados de la biblioteca pueden registrar libros nuevos.");
+        }
+
         BooksEntity bookEntity = new BooksEntity();
         bookEntity.setAuthor(bookDto.getAuthor());
         bookEntity.setTitle(bookDto.getTitle());
@@ -54,24 +68,61 @@ public class BooksService {
         return savedBookDto;
     }
 
-    public BooksDto updateBookStatus(String bookId, BooksDto bookDto) {
-        BooksEntity bookEntity = this.booksRepository.findById(bookId)
-                .orElseThrow(() -> new IllegalArgumentException("Libro con ID: " + bookId + " no encontrado"));
+    public BooksDto updateBookStatus(String bookId, BooksDto bookDto, UsersEntity user) {
+        try {
+            BooksEntity bookEntity = this.booksRepository.findById(bookId)
+                    .orElseThrow(() -> new IllegalArgumentException("Libro con ID: " + bookId + " no encontrado"));
 
-        bookEntity.setStatus(BookStatus.valueOf(bookDto.getStatus()));
-        bookEntity.setReservedBy(bookDto.getReservedBy());
-        bookEntity.setBorrowedBy(bookDto.getBorrowedBy());
+            try{
+                BookStatus.valueOf(bookDto.getStatus());
+            } catch (IllegalArgumentException e){
+                throw new BooksException("Ingresa un status valido. Los status validos son: DISPONIBLE, RESERVADO, PRESTADO.");
+            }
 
-        BooksEntity updatedBookEntity = this.booksRepository.save(bookEntity);
+            BooksDto currentBook = this.convertToDto(bookEntity);
 
-        BooksDto updatedBookDto = this.convertToDto(updatedBookEntity);
+            if (user.getRole() == Role.USER && currentBook.getStatus().equals(BookStatus.RESERVADO.toString())) {
+                throw new BooksException("Los usuarios no pueden cambiar el estado de un libro ya reservado.");
+            }
 
-        return updatedBookDto;
+            if (user.getRole() == Role.USER && !bookDto.getStatus().equals(BookStatus.RESERVADO.toString())) {
+                throw new BooksException("Los usuarios solo pueden reservar. Usa el status: RESERVADO.");
+            }
+
+            if (bookDto.getStatus().equals(BookStatus.RESERVADO.toString())) {
+                bookDto.setReservedBy(user.getEmail());
+            } else if (bookDto.getStatus().equals(BookStatus.PRESTADO.toString())) {
+                if (currentBook.getStatus().equals(BookStatus.RESERVADO.toString())) {
+                    bookDto.setReservedBy(currentBook.getReservedBy());
+                }
+                bookDto.setBorrowedBy(user.getEmail());
+            }
+
+            bookEntity.setStatus(BookStatus.valueOf(bookDto.getStatus()));
+            bookEntity.setReservedBy(bookDto.getReservedBy());
+            bookEntity.setBorrowedBy(bookDto.getBorrowedBy());
+
+            BooksEntity updatedBookEntity = this.booksRepository.save(bookEntity);
+
+            BooksDto updatedBookDto = this.convertToDto(updatedBookEntity);
+
+            return updatedBookDto;
+        } catch (BooksException e) {
+            throw new BooksException(e.getMessage());
+        }
     }
 
-    public void delete(String id) {
+    public void delete(String id, String userEmail) {
+        UsersEntity user = libraryRepository.findByEmail(userEmail)
+                .orElseThrow(() -> new UsernameNotFoundException("No se encontro al usuario con el correo: " + userEmail));
+
+        if (user.getRole() != Role.ADMIN) {
+            throw new BooksException("Solo los empleados de la biblioteca pueden eliminar libros.");
+        }
+
         BooksEntity booksEntity = this.booksRepository.findById(id)
-                .orElse(null);
+                .orElseThrow(() -> new IllegalArgumentException("No se encontro un libro con el ID: " + id));
+
         this.booksRepository.delete(booksEntity);
     }
